Refactor : replace course service
